Refinement of user interfaces
Added items such as the selection of what overall properties to report.
The ability to report pressure has been added, although this is of
limited use as it fluctuates the most out of all of the macroscopic
quantities.

Further developed README.md to specific what we want to achieve.

Changed a bit of CSS styling to help layout on mobile interfaces.