Added drawImage variation for better performance and GUI
molecule and element libraries now have images that are created when the
page is first loaded. This is probably suboptimal, and so I will think
about exporting them in the next update.

Using context.drawImage is significantly faster when there are several
thousanbd species on the page. Otherwise, the bottleneck remains at the
collision detection stage of affairs.

The reaction diagrams now also show an image of the molecules at
1px=20pm scale.