Code refactor and hydrocarbon combustion WIP setup
Made some changes to molecule type generation. This seeks to stream line
calculations a little so that coordinates don't need to be pre-centered.
add_entry() will now always zero the centre of mass.

Finally made a choice as to the combustion model that we will try. This
is the DRM22 extract from GriMech3.0, which emphasises reversibility and
works well across several conditions according to Zetterval et al., (2021).