Customised starting configurations and CSV export of data
Two main features have been introduced and earmarked as WIP:

(1) The use of copy-to-clipboard in order to save the ccurrent set of
initial configurations as a new HTML hyperlink with parameters.

(2) Manual exporting of the gas parameter graph data, and the gas
composition data as an CSV. Currently somewhat ugly and convoluted.